to many features were added. But main are the client cosmetics, new features in client such as fixed unread count, now you can send messages from client and so on. Also fixed unread counter in db and server. Soon Il fix all problems with client

+++ src/main/java/com/sunrise/cache/entity/Cache.java
@@ -1,6 +1,7 @@
 package com.sunrise.cache.entity;
 
 import java.time.Instant;
+import java.util.concurrent.atomic.AtomicInteger;
 
 import com.sunrise.orchestrator.type.ChatType;
 import com.sunrise.orchestrator.type.MessageType;
@@ -76,7 +77,7 @@ public class Cache {
         MessageType messageType,
         long senderId,
         String text,
-        long readCount,
+        AtomicInteger readCount,
         Instant sentAt,
         Instant updatedAt,
         Instant deletedAt,
