+++ src/main/java/com/sunrise/db/result/UserMessageResult.java
@@ -10,7 +10,7 @@ public interface UserMessageResult {
     Instant getProfileUpdatedAt();
     Instant getMemberUpdatedAt();
     String getText();
-    Long getReadCount();
+    Integer getReadCount();
     Instant getSentAt();
     Instant getUpdatedAt();
     Instant getDeletedAt();
